// src/order.dto.ts
export class Orderdto{
  termek: string;
  nev: string;
  kupon: string;

  szaml_orszag: string;
  szaml_iranyito: string;
  szaml_varos: string;
  szaml_cim: string;

  szall_orszag: string;
  szall_iranyito: string;
  szall_varos: string;
  szall_cim: string;
  
  bankkartya: string;
  lejarat: string;
  biztonsag: string;
}